ml-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
facenet-pytorch==2.5.3
faiss-cpu==1.7.4
Pillow==10.2.0
numpy==1.26.3
torch==2.1.2
torchvision==0.16.2
pydantic==2.5.3
